Change widgetHeader Test and create a test for booking info
