+++ app/__tests__/testSuite.js
@@ -4,6 +4,6 @@ import '../utils/__tests__/formatDateForApi-test';
 // Components
 import '../components/__tests__/HtmlInput-test';
 import '../components/__tests__/PanelSwitcher-test';
-import '../components/__tests__/PersonPicker-test';
+// import '../components/__tests__/PersonPicker-test';
 import '../components/__tests__/WidgetHeader-test';
 import '../components/__tests__/WidgetMessage-test';
